Fix int64 issue

+++ compile_and_profile.py
@@ -71,7 +71,7 @@ img_id = compile_model(
 txt_id = compile_model(
     model=onnx_txt_model, 
     device=target_device, 
-    input_specs={"text": ((1, 77), "int64")}
+    input_specs={"text": ((1, 77), "int32")}
 )
 
 print(f"Image compilation job ID: {img_id}")
